[#90] style: tidy up widget and imports
