+++ lib/features/live_writing/data/repositories/repositories.dart
@@ -1,1 +1,3 @@
 export 'confirm_post_repository_impl.dart';
+export 'live_writing_friend_repository_impl.dart';
+export 'live_writing_post_repository_impl.dart';
