+++ lib/features/live_writing/presentation/widgets/widgets.dart
@@ -1,4 +1,6 @@
 export 'confirm_post_content_input.dart';
 export 'confirm_post_form.dart';
 export 'confirm_post_title_input.dart';
+export 'live_writing_widget/friend_live_writing.dart';
+export 'live_writing_widget/my_live_writing.dart';
 export 'user_active_goal_dropdown.dart';
